Combined video and videoalpha acceptance tests. Resolved conflict between two steps with the same name.
