+++ lms/djangoapps/courseware/features/video.py
@@ -27,8 +27,30 @@ def view_video(_step):
     world.browser.visit(url)
 
 
+@step('the course has a VideoAlpha component')
+def view_videoalpha(step):
+    coursename = TEST_COURSE_NAME.replace(' ', '_')
+    i_am_registered_for_the_course(step, coursename)
+
+    # Make sure we have a videoalpha
+    add_videoalpha_to_course(coursename)
+    chapter_name = TEST_SECTION_NAME.replace(" ", "_")
+    section_name = chapter_name
+    url = django_url('/courses/edx/Test_Course/Test_Course/courseware/%s/%s' %
+                     (chapter_name, section_name))
+
+    world.browser.visit(url)
+
+
 def add_video_to_course(course):
     template_name = 'i4x://edx/templates/video/default'
     world.ItemFactory.create(parent_location=section_location(course),
                              template=template_name,
                              display_name='Video')
+
+
+def add_videoalpha_to_course(course):
+    template_name = 'i4x://edx/templates/videoalpha/default'
+    world.ItemFactory.create(parent_location=section_location(course),
+                             template=template_name,
+                             display_name='Video Alpha 1')
